dziala wszystko, tylko gamma wywala blad, ale to winna zaokraglen

+++ lab1/factorial/Factorial.h
@@ -5,6 +5,6 @@
 #ifndef JIMP_EXERCISES_FACTORIAL_H
 #define JIMP_EXERCISES_FACTORIAL_H
 
-double factorial(int value);
+int factorial(int value);
 
 #endif //JIMP_EXERCISES_FACTORIAL_H
